Move reader functions out of the feed class
This is a design that simplifies the feed class and
prevents confusing results from filtered feeds.
