+++ tests/test_feed.py
@@ -153,84 +153,6 @@ def test_config_must_be_dag():
         assert 'Config must be a DAG' in repr(e)
 
 
-def test_service_ids_by_date():
-    path = fixture('amazon-2017-08-06.zip')
-
-    service_ids_by_date = ptg.read_service_ids_by_date(path)
-    feed = ptg.feed(path)
-    raw_feed = ptg.raw_feed(path)
-
-    assert service_ids_by_date == feed.service_ids_by_date
-    assert service_ids_by_date == raw_feed.service_ids_by_date
-
-    assert service_ids_by_date == {
-        datetime.date(2017, 8, 1): frozenset({'1', '0'}),
-        datetime.date(2017, 8, 2): frozenset({'1', '0'}),
-        datetime.date(2017, 8, 3): frozenset({'1', '0'}),
-        datetime.date(2017, 8, 4): frozenset({'1', '0'}),
-        datetime.date(2017, 8, 5): frozenset({'1'}),
-        datetime.date(2017, 8, 7): frozenset({'1', '0'})
-    }
-
-
-def test_dates_by_service_ids():
-    path = fixture('amazon-2017-08-06.zip')
-
-    dates_by_service_ids = ptg.read_dates_by_service_ids(path)
-    feed = ptg.feed(path)
-    raw_feed = ptg.raw_feed(path)
-
-    assert dates_by_service_ids == feed.dates_by_service_ids
-    assert dates_by_service_ids == raw_feed.dates_by_service_ids
-
-    assert dates_by_service_ids == {
-        frozenset({'1'}): {
-            datetime.date(2017, 8, 5),
-        },
-        frozenset({'1', '0'}): {
-            datetime.date(2017, 8, 1),
-            datetime.date(2017, 8, 2),
-            datetime.date(2017, 8, 3),
-            datetime.date(2017, 8, 4),
-            datetime.date(2017, 8, 7),
-        }
-    }
-
-
-def test_trip_counts_by_date():
-    path = fixture('amazon-2017-08-06.zip')
-
-    trip_counts_by_date = ptg.read_trip_counts_by_date(path)
-    feed = ptg.feed(path)
-    raw_feed = ptg.raw_feed(path)
-
-    assert trip_counts_by_date == feed.trip_counts_by_date
-    assert trip_counts_by_date == raw_feed.trip_counts_by_date
-
-    assert trip_counts_by_date == {
-        datetime.date(2017, 8, 1): 442,
-        datetime.date(2017, 8, 2): 442,
-        datetime.date(2017, 8, 3): 442,
-        datetime.date(2017, 8, 4): 442,
-        datetime.date(2017, 8, 5): 1,
-        datetime.date(2017, 8, 7): 442,
-    }
-
-
-@pytest.mark.parametrize('path', [
-    fixture('empty.zip'),
-    fixture('amazon-2017-08-06.zip'),
-    fixture('caltrain-2017-07-24.zip'),
-])
-def test_structure(path):
-    feed = ptg.feed(path)
-    for date, service_ids in feed.service_ids_by_date.items():
-        member_dates = feed.dates_by_service_ids[service_ids]
-        for member in member_dates:
-            assert feed.service_ids_by_date[member] == service_ids
-            assert feed.trip_counts_by_date[member] > 0
-
-
 @pytest.mark.parametrize('path', [
     fixture('amazon-2017-08-06.zip'),
     fixture('caltrain-2017-07-24.zip'),
